Starting the process of converting the API to use WordPress driven queries instead of normalized queries
Long story short. For now I am just going with standard WordPress queries, which are about 10x slower than normalized queries (~200ms vs ~30ms) since the process of normalizing adds another level of complexity and puts the data in two spots.

If this site ever gains traction we will definitely want to move back to the normalization process and build our a different backend for adding videos to the DB. For now though I am taking advantage of the built in WordPress multisite functionality and admin area which I really don't want to have to rebuild in another framework atm.